requirements.txt
#Django>=5.0,<6.0
django-allauth[socialaccount]>=0.61
django-crispy-forms>=2.1
crispy-bootstrap5>=2024.2
django-environ>=0.11
Pillow>=10.0
whitenoise>=6.6
requests>=2.31

# Dev only
django-debug-toolbar>=4.3